# Allow custom key comparator