--- a/compaction.go
+++ b/compaction.go
@@ -30,12 +30,15 @@ import (
 )
 
 type keyRange struct {
-	left  []byte
-	right []byte
-	inf   bool
+	left          []byte
+	right         []byte
+	inf           bool
+	keyComparator y.KeyComparator
 }
 
-var infRange = keyRange{inf: true}
+func infRange(comp y.KeyComparator) keyRange {
+	return keyRange{inf: true, keyComparator: comp}
+}
 
 func (r keyRange) String() string {
 	return fmt.Sprintf("[left=%x, right=%x, inf=%v]", r.left, r.right, r.inf)
@@ -53,20 +56,20 @@ func (r keyRange) overlapsWith(dst keyRange) bool {
 	}
 
 	// If my left is greater than dst right, we have no overlap.
-	if y.CompareKeys(r.left, dst.right) > 0 {
+	if r.keyComparator.CompareKeys(r.left, dst.right) > 0 {
 		return false
 	}
 	// If my right is less than dst left, we have no overlap.
-	if y.CompareKeys(r.right, dst.left) < 0 {
+	if r.keyComparator.CompareKeys(r.right, dst.left) < 0 {
 		return false
 	}
 	// We have overlap.
 	return true
 }
 
-func getKeyRange(tables []*table.Table) keyRange {
+func getKeyRange(tables []*table.Table, comp y.KeyComparator) keyRange {
 	if len(tables) == 0 {
-		return keyRange{}
+		return keyRange{keyComparator: comp}
 	}
 	smallest := tables[0].Smallest()
 	biggest := tables[0].Biggest()
@@ -79,8 +82,9 @@ func getKeyRange(tables []*table.Table) keyRange {
 		}
 	}
 	return keyRange{
-		left:  y.KeyWithTs(y.ParseKey(smallest), math.MaxUint64),
-		right: y.KeyWithTs(y.ParseKey(biggest), 0),
+		left:          y.KeyWithTs(y.ParseKey(smallest), math.MaxUint64),
+		right:         y.KeyWithTs(y.ParseKey(biggest), 0),
+		keyComparator: comp,
 	}
 }