--- a/db.go
+++ b/db.go
@@ -273,7 +273,7 @@ func Open(opt Options) (db *DB, err error) {
 	db.calculateSize()
 	db.closers.updateSize = y.NewCloser(1)
 	go db.updateSize(db.closers.updateSize)
-	db.mt = skl.NewSkiplist(arenaSize(opt))
+	db.mt = skl.NewSkiplist(arenaSize(opt), opt.KeyComparator)
 
 	// newLevelsController potentially loads files in directory.
 	if db.lc, err = newLevelsController(db, &manifest); err != nil {
@@ -799,7 +799,7 @@ func (db *DB) ensureRoomForWrite() error {
 			db.mt.MemSize(), len(db.flushChan))
 		// We manage to push this task. Let's modify imm.
 		db.imm = append(db.imm, db.mt)
-		db.mt = skl.NewSkiplist(arenaSize(db.opt))
+		db.mt = skl.NewSkiplist(arenaSize(db.opt), opt.KeyComparator)
 		// New memtable is empty. We certainly have room.
 		return nil
 	default:
@@ -1344,7 +1344,7 @@ func (db *DB) dropAll() (func(), error) {
 		mt.DecrRef()
 	}
 	db.imm = db.imm[:0]
-	db.mt = skl.NewSkiplist(arenaSize(db.opt)) // Set it up for future writes.
+	db.mt = skl.NewSkiplist(arenaSize(db.opt), opt.KeyComparator) // Set it up for future writes.
 
 	num, err := db.lc.dropTree()
 	if err != nil {
@@ -1400,7 +1400,7 @@ func (db *DB) DropPrefix(prefix []byte) error {
 		memtable.DecrRef()
 	}
 	db.imm = db.imm[:0]
-	db.mt = skl.NewSkiplist(arenaSize(db.opt))
+	db.mt = skl.NewSkiplist(arenaSize(db.opt), opt.KeyComparator)
 
 	// Drop prefixes from the levels.
 	if err := db.lc.dropPrefix(prefix); err != nil {