--- a/db.go
+++ b/db.go
@@ -799,7 +799,7 @@ func (db *DB) ensureRoomForWrite() error {
 			db.mt.MemSize(), len(db.flushChan))
 		// We manage to push this task. Let's modify imm.
 		db.imm = append(db.imm, db.mt)
-		db.mt = skl.NewSkiplist(arenaSize(db.opt), opt.KeyComparator)
+		db.mt = skl.NewSkiplist(arenaSize(db.opt), db.opt.KeyComparator)
 		// New memtable is empty. We certainly have room.
 		return nil
 	default:
@@ -881,7 +881,7 @@ func (db *DB) handleFlushTask(ft flushTask) error {
 		db.elog.Errorf("ERROR while syncing level directory: %v", dirSyncErr)
 	}
 
-	tbl, err := table.OpenTable(fd, db.opt.TableLoadingMode, nil)
+	tbl, err := table.OpenTable(fd, db.opt.TableLoadingMode, db.opt.KeyComparator, nil)
 	if err != nil {
 		db.elog.Printf("ERROR while opening table: %v", err)
 		return err
@@ -1344,7 +1344,7 @@ func (db *DB) dropAll() (func(), error) {
 		mt.DecrRef()
 	}
 	db.imm = db.imm[:0]
-	db.mt = skl.NewSkiplist(arenaSize(db.opt), opt.KeyComparator) // Set it up for future writes.
+	db.mt = skl.NewSkiplist(arenaSize(db.opt), db.opt.KeyComparator) // Set it up for future writes.
 
 	num, err := db.lc.dropTree()
 	if err != nil {
@@ -1400,7 +1400,7 @@ func (db *DB) DropPrefix(prefix []byte) error {
 		memtable.DecrRef()
 	}
 	db.imm = db.imm[:0]
-	db.mt = skl.NewSkiplist(arenaSize(db.opt), opt.KeyComparator)
+	db.mt = skl.NewSkiplist(arenaSize(db.opt), db.opt.KeyComparator)
 
 	// Drop prefixes from the levels.
 	if err := db.lc.dropPrefix(prefix); err != nil {