--- a/level_handler.go
+++ b/level_handler.go
@@ -69,7 +69,7 @@ func (s *levelHandler) initTables(tables []*table.Table) {
 	} else {
 		// Sort tables by keys.
 		sort.Slice(s.tables, func(i, j int) bool {
-			return y.CompareKeys(s.tables[i].Smallest(), s.tables[j].Smallest()) < 0
+			return s.db.opt.KeyComparator.CompareKeys(s.tables[i].Smallest(), s.tables[j].Smallest()) < 0
 		})
 	}
 }
@@ -132,7 +132,7 @@ func (s *levelHandler) replaceTables(toDel, toAdd []*table.Table) error {
 	// Assign tables.
 	s.tables = newTables
 	sort.Slice(s.tables, func(i, j int) bool {
-		return y.CompareKeys(s.tables[i].Smallest(), s.tables[j].Smallest()) < 0
+		return s.db.opt.KeyComparator.CompareKeys(s.tables[i].Smallest(), s.tables[j].Smallest()) < 0
 	})
 	s.Unlock() // s.Unlock before we DecrRef tables -- that can be slow.
 	return decrRefs(toDel)
@@ -215,7 +215,7 @@ func (s *levelHandler) getTableForKey(key []byte) ([]*table.Table, func() error)
 	}
 	// For level >= 1, we can do a binary search as key range does not overlap.
 	idx := sort.Search(len(s.tables), func(i int) bool {
-		return y.CompareKeys(s.tables[i].Biggest(), key) >= 0
+		return s.db.opt.KeyComparator.CompareKeys(s.tables[i].Biggest(), key) >= 0
 	})
 	if idx >= len(s.tables) {
 		// Given key is strictly > than every element we have.
@@ -290,10 +290,10 @@ func (s *levelHandler) overlappingTables(_ levelHandlerRLocked, kr keyRange) (in
 		return 0, 0
 	}
 	left := sort.Search(len(s.tables), func(i int) bool {
-		return y.CompareKeys(kr.left, s.tables[i].Biggest()) <= 0
+		return s.db.opt.KeyComparator.CompareKeys(kr.left, s.tables[i].Biggest()) <= 0
 	})
 	right := sort.Search(len(s.tables), func(i int) bool {
-		return y.CompareKeys(kr.right, s.tables[i].Smallest()) < 0
+		return s.db.opt.KeyComparator.CompareKeys(kr.right, s.tables[i].Smallest()) < 0
 	})
 	return left, right
 }