--- a/levels.go
+++ b/levels.go
@@ -390,7 +390,7 @@ func (s *levelsController) pickCompactLevels() (prios []compactionPriority) {
 	// addLevel0Table uses.
 
 	// cstatus is checked to see if level 0's tables are already being compacted
-	if !s.cstatus.overlapsWith(0, infRange) && s.isLevel0Compactable() {
+	if !s.cstatus.overlapsWith(0, infRange(s.kv.opt.KeyComparator)) && s.isLevel0Compactable() {
 		pri := compactionPriority{
 			level: 0,
 			score: float64(s.levels[0].numTables()) / float64(s.kv.opt.NumLevelZeroTables),
@@ -424,7 +424,7 @@ func (s *levelsController) compactBuildTables(
 
 	var hasOverlap bool
 	{
-		kr := getKeyRange(cd.top)
+		kr := getKeyRange(cd.top, s.kv.opt.KeyComparator)
 		for i, lh := range s.levels {
 			if i <= lev { // Skip upper levels.
 				continue
@@ -472,7 +472,7 @@ func (s *levelsController) compactBuildTables(
 		valid = append(valid, table)
 	}
 	iters = append(iters, table.NewConcatIterator(valid, false))
-	it := y.NewMergeIterator(iters, false)
+	it := y.NewMergeIterator(iters, s.kv.opt.KeyComparator, false)
 	defer it.Close() // Important to close the iterator to do ref counting.
 
 	it.Rewind()
@@ -675,9 +675,9 @@ func (s *levelsController) fillTablesL0(cd *compactDef) bool {
 	if len(cd.top) == 0 {
 		return false
 	}
-	cd.thisRange = infRange
+	cd.thisRange = infRange(s.kv.opt.KeyComparator)
 
-	kr := getKeyRange(cd.top)
+	kr := getKeyRange(cd.top, s.kv.opt.KeyComparator)
 	left, right := cd.nextLevel.overlappingTables(levelHandlerRLocked{}, kr)
 	cd.bot = make([]*table.Table, right-left)
 	copy(cd.bot, cd.nextLevel.tables[left:right])
@@ -685,7 +685,7 @@ func (s *levelsController) fillTablesL0(cd *compactDef) bool {
 	if len(cd.bot) == 0 {
 		cd.nextRange = kr
 	} else {
-		cd.nextRange = getKeyRange(cd.bot)
+		cd.nextRange = getKeyRange(cd.bot, s.kv.opt.KeyComparator)
 	}
 
 	if !s.cstatus.compareAndAdd(thisAndNextLevelRLocked{}, *cd) {
@@ -717,7 +717,8 @@ func (s *levelsController) fillTables(cd *compactDef) bool {
 			// We pick all the versions of the smallest and the biggest key.
 			left: y.KeyWithTs(y.ParseKey(t.Smallest()), math.MaxUint64),
 			// Note that version zero would be the rightmost key.
-			right: y.KeyWithTs(y.ParseKey(t.Biggest()), 0),
+			right:         y.KeyWithTs(y.ParseKey(t.Biggest()), 0),
+			keyComparator: s.kv.opt.KeyComparator,
 		}
 		if s.cstatus.overlapsWith(cd.thisLevel.level, cd.thisRange) {
 			continue
@@ -736,7 +737,7 @@ func (s *levelsController) fillTables(cd *compactDef) bool {
 			}
 			return true
 		}
-		cd.nextRange = getKeyRange(cd.bot)
+		cd.nextRange = getKeyRange(cd.bot, s.kv.opt.KeyComparator)
 
 		if s.cstatus.overlapsWith(cd.nextLevel.level, cd.nextRange) {
 			continue