--- a/stream_writer.go
+++ b/stream_writer.go
@@ -262,7 +262,7 @@ func (w *sortedWriter) createTable(data []byte) error {
 	if _, err := fd.Write(data); err != nil {
 		return err
 	}
-	tbl, err := table.OpenTable(fd, w.db.opt.TableLoadingMode, nil)
+	tbl, err := table.OpenTable(fd, w.db.opt.TableLoadingMode, w.db.opt.KeyComparator, nil)
 	if err != nil {
 		return err
 	}