--- a/util.go
+++ b/util.go
@@ -76,14 +76,14 @@ func (s *levelHandler) validate() error {
 			return errors.Errorf("Level %d, j=%d numTables=%d", s.level, j, numTables)
 		}
 
-		if y.CompareKeys(s.tables[j-1].Biggest(), s.tables[j].Smallest()) >= 0 {
+		if s.db.opt.KeyComparator.CompareKeys(s.tables[j-1].Biggest(), s.tables[j].Smallest()) >= 0 {
 			return errors.Errorf(
 				"Inter: Biggest(j-1) \n%s\n vs Smallest(j): \n%s\n: level=%d j=%d numTables=%d",
 				hex.Dump(s.tables[j-1].Biggest()), hex.Dump(s.tables[j].Smallest()),
 				s.level, j, numTables)
 		}
 
-		if y.CompareKeys(s.tables[j].Smallest(), s.tables[j].Biggest()) > 0 {
+		if s.db.opt.KeyComparator.CompareKeys(s.tables[j].Smallest(), s.tables[j].Biggest()) > 0 {
 			return errors.Errorf(
 				"Intra: %q vs %q: level=%d j=%d numTables=%d",
 				s.tables[j].Smallest(), s.tables[j].Biggest(), s.level, j, numTables)