This PR adds a global variable in the `y` package called `CompareKey` that allows users to change the behaviour of the key comparator used to decide the key ordering.
I first wanted to provide that as an option but given the locations where the original `CompareKey` function is used, it would have been difficult to pass the option everywhere.
This solution seems to be the cheapest in terms of changes, the tradeoff being allowing only one key comparator function at a time for all Badger instances.
Let me know if you prefer an other way.

Fixes #776

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/dgraph-io/badger/815)
<!-- Reviewable:end -->