@achille-roussel 2019-05-24
Using a global comparator means that a single program cannot manipulate multiple databases that use different key ordering. I'm not an expert on the codebase but would it be preferable/possible for the comparison function to be local to a database (maybe passing it to [Options](https://godoc.org/github.com/dgraph-io/badger#Options))?