@martinmr 2019-05-25
@achille-roussel Yes that's a good idea. We do something similar already for the Logger, which is db specific.

@asdine  Can you implement this suggestion. There should be a field in Options and if it's not set badger will use the default comparison function. That way multiple dbs can each have a different comparison func.