@campoy 2019-05-28
If we implement the change and provide a way to configure what key comparison we'll use we gotta make sure the performance doesn't get affected by this.

This little function is going to be called millions of times, so a slight change in it could have huge impact.