@achille-roussel 2019-05-28
It’s very likely that there could be a performance regression.

@campoy do you have a sense of how much CPU time we’re OK to spare to introduce this feature?