@asdine 2019-06-01
I have some weird fails on AppVeyor that I can't reproduce locally.
Before diving further, could you guys tell me if the solution is good enough?

Here is a benchmark comparision between the master branch and this one:

```
benchmark                                          old ns/op     new ns/op     delta
BenchmarkIteratePrefixSingleKey/Key_lookups-12     72543         72392         -0.21%
BenchmarkReadWrite/0.1,0064-12                     3542          3340          -5.70%
BenchmarkReadWrite/0.2,0064-12                     3603          3543          -1.67%
BenchmarkReadWrite/0.5,0064-12                     4328          4213          -2.66%
BenchmarkReadWrite/1.0,0064-12                     5193          5428          +4.53%
BenchmarkReadWrite/0.1,0128-12                     3359          3379          +0.60%
BenchmarkReadWrite/0.2,0128-12                     3451          3597          +4.23%
BenchmarkReadWrite/0.5,0128-12                     4339          4311          -0.65%
BenchmarkReadWrite/1.0,0128-12                     5241          5535          +5.61%
BenchmarkReadWrite/0.1,0256-12                     3436          3510          +2.15%
BenchmarkReadWrite/0.2,0256-12                     3747          3759          +0.32%
BenchmarkReadWrite/0.5,0256-12                     4518          4642          +2.74%
BenchmarkReadWrite/1.0,0256-12                     5547          5794          +4.45%
BenchmarkReadWrite/0.1,0512-12                     4423          3931          -11.12%
BenchmarkReadWrite/0.2,0512-12                     3999          4123          +3.10%
BenchmarkReadWrite/0.5,0512-12                     4857          5112          +5.25%
BenchmarkReadWrite/1.0,0512-12                     6284          6216          -1.08%
BenchmarkReadWrite/0.1,1024-12                     4615          5120          +10.94%
BenchmarkReadWrite/0.2,1024-12                     5815          5749          -1.13%
BenchmarkReadWrite/0.5,1024-12                     5937          5656          -4.73%
BenchmarkReadWrite/1.0,1024-12                     6927          8267          +19.34%
BenchmarkReadWrite/0.1,2048-12                     6193          6427          +3.78%
BenchmarkReadWrite/0.2,2048-12                     6740          7715          +14.47%
BenchmarkReadWrite/0.5,2048-12                     7316          7993          +9.25%
BenchmarkReadWrite/1.0,2048-12                     11286         10071         -10.77%
BenchmarkReadWrite/0.1,4096-12                     8813          9165          +3.99%
BenchmarkReadWrite/0.2,4096-12                     8562          11810         +37.94%
BenchmarkReadWrite/0.5,4096-12                     24010         12724         -47.01%
BenchmarkReadWrite/1.0,4096-12                     13777         17054         +23.79%
BenchmarkReadWrite/0.1,8192-12                     20057         19375         -3.40%
BenchmarkReadWrite/0.2,8192-12                     16960         15810         -6.78%
BenchmarkReadWrite/0.5,8192-12                     49764         43662         -12.26%
BenchmarkReadWrite/1.0,8192-12                     53512         49405         -7.67%
BenchmarkReadWrite/0.1,16384-12                    52576         26682         -49.25%
BenchmarkReadWrite/0.2,16384-12                    55470         54218         -2.26%
BenchmarkReadWrite/0.5,16384-12                    95961         88120         -8.17%
BenchmarkReadWrite/1.0,16384-12                    50505         60801         +20.39%

benchmark                                          old allocs     new allocs     delta
BenchmarkIteratePrefixSingleKey/Key_lookups-12     556            564            +1.44%
BenchmarkReadWrite/0.1,0064-12                     26             26             +0.00%
BenchmarkReadWrite/0.2,0064-12                     28             28             +0.00%
BenchmarkReadWrite/0.5,0064-12                     33             33             +0.00%
BenchmarkReadWrite/1.0,0064-12                     41             41             +0.00%
BenchmarkReadWrite/0.1,0128-12                     26             26             +0.00%
BenchmarkReadWrite/0.2,0128-12                     28             28             +0.00%
BenchmarkReadWrite/0.5,0128-12                     33             33             +0.00%
BenchmarkReadWrite/1.0,0128-12                     41             41             +0.00%
BenchmarkReadWrite/0.1,0256-12                     26             26             +0.00%
BenchmarkReadWrite/0.2,0256-12                     28             28             +0.00%
BenchmarkReadWrite/0.5,0256-12                     32             33             +3.12%
BenchmarkReadWrite/1.0,0256-12                     41             41             +0.00%
BenchmarkReadWrite/0.1,0512-12                     26             26             +0.00%
BenchmarkReadWrite/0.2,0512-12                     28             28             +0.00%
BenchmarkReadWrite/0.5,0512-12                     32             32             +0.00%
BenchmarkReadWrite/1.0,0512-12                     41             41             +0.00%
BenchmarkReadWrite/0.1,1024-12                     26             26             +0.00%
BenchmarkReadWrite/0.2,1024-12                     28             28             +0.00%
BenchmarkReadWrite/0.5,1024-12                     32             32             +0.00%
BenchmarkReadWrite/1.0,1024-12                     41             41             +0.00%
BenchmarkReadWrite/0.1,2048-12                     26             26             +0.00%
BenchmarkReadWrite/0.2,2048-12                     28             28             +0.00%
BenchmarkReadWrite/0.5,2048-12                     32             33             +3.12%
BenchmarkReadWrite/1.0,2048-12                     41             41             +0.00%
BenchmarkReadWrite/0.1,4096-12                     26             26             +0.00%
BenchmarkReadWrite/0.2,4096-12                     28             28             +0.00%
BenchmarkReadWrite/0.5,4096-12                     33             32             -3.03%
BenchmarkReadWrite/1.0,4096-12                     41             41             +0.00%
BenchmarkReadWrite/0.1,8192-12                     26             26             +0.00%
BenchmarkReadWrite/0.2,8192-12                     28             28             +0.00%
BenchmarkReadWrite/0.5,8192-12                     33             33             +0.00%
BenchmarkReadWrite/1.0,8192-12                     41             41             +0.00%
BenchmarkReadWrite/0.1,16384-12                    26             26             +0.00%
BenchmarkReadWrite/0.2,16384-12                    28             28             +0.00%
BenchmarkReadWrite/0.5,16384-12                    33             33             +0.00%
BenchmarkReadWrite/1.0,16384-12                    41             41             +0.00%

benchmark                                          old bytes     new bytes     delta
BenchmarkIteratePrefixSingleKey/Key_lookups-12     30549         26661         -12.73%
BenchmarkReadWrite/0.1,0064-12                     972           972           +0.00%
BenchmarkReadWrite/0.2,0064-12                     1032          1032          +0.00%
BenchmarkReadWrite/0.5,0064-12                     1213          1213          +0.00%
BenchmarkReadWrite/1.0,0064-12                     1512          1512          +0.00%
BenchmarkReadWrite/0.1,0128-12                     1106          1106          +0.00%
BenchmarkReadWrite/0.2,0128-12                     1172          1172          +0.00%
BenchmarkReadWrite/0.5,0128-12                     1373          1373          +0.00%
BenchmarkReadWrite/1.0,0128-12                     1704          1704          +0.00%
BenchmarkReadWrite/0.1,0256-12                     1376          1375          -0.07%
BenchmarkReadWrite/0.2,0256-12                     1455          1455          +0.00%
BenchmarkReadWrite/0.5,0256-12                     1691          1694          +0.18%
BenchmarkReadWrite/1.0,0256-12                     2088          2088          +0.00%
BenchmarkReadWrite/0.1,0512-12                     1913          1912          -0.05%
BenchmarkReadWrite/0.2,0512-12                     2018          2017          -0.05%
BenchmarkReadWrite/0.5,0512-12                     2332          2331          -0.04%
BenchmarkReadWrite/1.0,0512-12                     2856          2856          +0.00%
BenchmarkReadWrite/0.1,1024-12                     2989          2988          -0.03%
BenchmarkReadWrite/0.2,1024-12                     3144          3145          +0.03%
BenchmarkReadWrite/0.5,1024-12                     3612          3611          -0.03%
BenchmarkReadWrite/1.0,1024-12                     4392          4392          +0.00%
BenchmarkReadWrite/0.1,2048-12                     5280          5277          -0.06%
BenchmarkReadWrite/0.2,2048-12                     5554          5550          -0.07%
BenchmarkReadWrite/0.5,2048-12                     6358          6370          +0.19%
BenchmarkReadWrite/1.0,2048-12                     7721          7721          +0.00%
BenchmarkReadWrite/0.1,4096-12                     10146         10140         -0.06%
BenchmarkReadWrite/0.2,4096-12                     10670         10673         +0.03%
BenchmarkReadWrite/0.5,4096-12                     12255         12249         -0.05%
BenchmarkReadWrite/1.0,4096-12                     14890         14890         +0.00%
BenchmarkReadWrite/0.1,8192-12                     19310         19323         +0.07%
BenchmarkReadWrite/0.2,8192-12                     20285         20289         +0.02%
BenchmarkReadWrite/0.5,8192-12                     23274         23261         -0.06%
BenchmarkReadWrite/1.0,8192-12                     28203         28203         +0.00%
BenchmarkReadWrite/0.1,16384-12                    37376         37351         -0.07%
BenchmarkReadWrite/0.2,16384-12                    39236         39229         -0.02%
BenchmarkReadWrite/0.5,16384-12                    44895         44886         -0.02%
BenchmarkReadWrite/1.0,16384-12                    54319         54319         +0.00%
```